+++ simple_history/admin.py
@@ -16,7 +16,9 @@ from django.utils.html import mark_safe
 from django.utils.text import capfirst
 from django.utils.translation import gettext as _
 
-from .manager import HistoryManager
+from .manager import HistoricalQuerySet, HistoryManager
+from .models import HistoricalChanges
+from .template_utils import HistoricalRecordContextHelper
 from .utils import get_history_manager_for_model, get_history_model_for_model
 
 SIMPLE_HISTORY_EDIT = getattr(settings, "SIMPLE_HISTORY_EDIT", False)
@@ -76,14 +78,16 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
                 for record in historical_records:
                     setattr(record, history_list_entry, value_for_entry(record))
 
+        self.set_history_delta_changes(request, historical_records)
+
         content_type = self.content_type_model_cls.objects.get_for_model(
             get_user_model()
         )
-
         admin_user_view = "admin:{}_{}_change".format(
             content_type.app_label,
             content_type.model,
         )
+
         context = {
             "title": self.history_view_title(request, obj),
             "object_history_list_template": self.object_history_list_template,
@@ -117,10 +121,12 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
         :param pk_name: The name of the original model's primary key field.
         :param object_id: The primary key of the object whose history is listed.
         """
-        qs = history_manager.filter(**{pk_name: object_id})
+        qs: HistoricalQuerySet = history_manager.filter(**{pk_name: object_id})
         if not isinstance(history_manager.model.history_user, property):
             # Only select_related when history_user is a ForeignKey (not a property)
             qs = qs.select_related("history_user")
+        # Prefetch related objects to reduce the number of DB queries when diffing
+        qs = qs._select_related_history_tracked_objs()
         return qs
 
     def get_history_list_display(self, request) -> Sequence[str]:
@@ -130,6 +136,44 @@ class SimpleHistoryAdmin(admin.ModelAdmin):
         or the history model, or methods on the admin class.
         """
         return self.history_list_display
+
+    def get_historical_record_context_helper(
+        self, request, historical_record: HistoricalChanges
+    ) -> HistoricalRecordContextHelper:
+        """
+        Return an instance of ``HistoricalRecordContextHelper`` for formatting
+        the template context for ``historical_record``.
+        """
+        return HistoricalRecordContextHelper(self.model, historical_record)
+
+    def set_history_delta_changes(
+        self,
+        request,
+        historical_records: Sequence[HistoricalChanges],
+        foreign_keys_are_objs=True,
+    ):
+        """
+        Add a ``history_delta_changes`` attribute to all historical records
+        except the first (oldest) one.
+
+        :param request:
+        :param historical_records:
+        :param foreign_keys_are_objs: Passed to ``diff_against()`` when calculating
+               the deltas; see its docstring for details.
+        """
+        previous = None
+        for current in historical_records:
+            if previous is None:
+                previous = current
+                continue
+            # Related objects should have been prefetched in `get_history_queryset()`
+            delta = previous.diff_against(
+                current, foreign_keys_are_objs=foreign_keys_are_objs
+            )
+            helper = self.get_historical_record_context_helper(request, previous)
+            previous.history_delta_changes = helper.context_for_delta_changes(delta)
+
+            previous = current
 
     def history_view_title(self, request, obj):
         if self.revert_disabled(request, obj) and not SIMPLE_HISTORY_EDIT: